Implement record synchronization.
This commit introduces `ODRecordSynchronizer` which is used to make operations
on behalf of `ODRecordStorage` when syncing is needed.

Other classes are updated to set up the synchronizer. Storage apply updates
when relevant methods are being called.

Include minor fixes including `-[ODRecord copy]` which does not create
`NSMutableDictionary` for internal use.

Current limitations:

*   Pending saves that are not synched to the server will be lost after
    app restart.
*   Record storage will not automatically retry failed operations, but
    it will try again if there is new changes appended.

refs #78