Refactor completion logic for ODOperation. #100
The `ODOperation` completion logic is refactored so that it is easier
for subclass to implement result handling. The behavior is backward-compatible,
meaning existing subclasse implementations continue to work.

Previously, subclass of `ODOperation` has to set the `completionBlock`
property to handle the error and response by retrieving those from
a property of `ODOperation`.

The logic is changed so that subclass should implement `-handleRequestError:`
and `-handleResponse:`, which are called by `ODOperation` when the operation
completes.

`ODFetchRecordsOperation` has been modified to make use of the new logic
as an example for future implementation.

refs #100