Implements record query and sorting.
First of all, `ODRecord` is now key-value compliant on the record attributes,
which is required to support query and sorting.

All backing stores are modified to support query and sorting. The SQLite
also supports these features, but it achieves this in the same way
as the memory store, which is by calling `-[NSPredicate evaluateWithObject:]`
and -[NSArray sortUsingDescriptors:]`. The SQLite may support more
efficient query (by using database index) in the future.

refs #77