Set operation error in case of partial error
Previously when ModifyRecordsOperation has problem saving one of the
records to be saved, the operation error returned will be nil. This commit
change the behaviour to returning a partial error instead.

Also ensured in this commit is that when saving a record, the record
(together with any updated fields) are returned through completion block.

refs #189