Clear up websocket state when socket closed
When connecting to skygear-server in docker and kill the docker container, it is expected that connect did fail get called during the first reconnect attempt, but instead socket did close get called. This lead to a wrong state of SKYPubsubClient, which would no longer connect to server anymore.

Clearing the state would make the websocket return to the reconnect flow, and when server restarted, the websocket would reconnect normally. However, socket did close would still be called multiple times.